+++ packages/inversiland/src/middlewares/debugMiddleware.ts
@@ -7,7 +7,7 @@ function debugMiddleware(planAndResolve: interfaces.Next): interfaces.Next {
   return (args: interfaces.NextArgs) => {
     const nextContextInterceptor = args.contextInterceptor;
     args.contextInterceptor = (context: interfaces.Context) => {
-      if (inversilandOptions.debug) {
+      if (inversilandOptions.logLevel === "debug") {
         console.log(
           messagesMap.providerRequested(
             args.serviceIdentifier,