+++ packages/inversiland/src/types/index.ts
@@ -4,6 +4,7 @@ import {
   Factory,
   FactoryWrapper,
 } from "./factories";
+import InversilandLogLevel from "./InversilandLogLevel";
 import InversilandOptions from "./InversilandOptions";
 import InversilandState from "./InversilandState";
 import { DynamicModule, NewableModule } from "./Module";
@@ -13,6 +14,7 @@ import Provider from "./Provider";
 export {
   Newable,
   Provider,
+  InversilandLogLevel,
   InversilandOptions,
   InversilandState,
   NewableModule,