+++ packages/inversiland/test/middlewares/debugMiddleware.spec.ts
@@ -5,11 +5,11 @@ import inversilandOptions from "../../src/inversiland/inversilandOptions";
 import messagesMap from "../../src/messages/messagesMap";
 
 describe("debugMiddleware", () => {
-  it("Should log the correct message.", () => {
+  it("Should log the correct message when logLevel = 'debug'.", () => {
     @injectable()
     class TestService {}
 
-    inversilandOptions.debug = true;
+    inversilandOptions.logLevel = "debug";
 
     const consoleLogMock = jest.spyOn(console, "log").mockImplementation();
     const container = new Container();
@@ -18,16 +18,16 @@ describe("debugMiddleware", () => {
     container.bind(TestService).toSelf();
     container.get(TestService);
 
-    expect(consoleLogMock).toBeCalledWith(
+    expect(consoleLogMock).toHaveBeenCalledWith(
       messagesMap.providerRequested(TestService, container.id)
     );
   });
 
-  it("Should not log the message when debug is false.", () => {
+  it("Should not log the message when logLevel is 'info'.", () => {
     @injectable()
     class TestService {}
 
-    inversilandOptions.debug = false;
+    inversilandOptions.logLevel = "info";
 
     const consoleLogMock = jest.spyOn(console, "log").mockImplementation();
     const container = new Container();
@@ -36,19 +36,37 @@ describe("debugMiddleware", () => {
     container.bind(TestService).toSelf();
     container.get(TestService);
 
-    expect(consoleLogMock).not.toBeCalledWith(
+    expect(consoleLogMock).not.toHaveBeenCalledWith(
       messagesMap.providerRequested(TestService, container.id)
     );
   });
 
-  it("Should log a message for each provider resolved.", () => {
+  it("Should not log the message when logLevel is 'none'.", () => {
+    @injectable()
+    class TestService {}
+
+    inversilandOptions.logLevel = "none";
+
+    const consoleLogMock = jest.spyOn(console, "log").mockImplementation();
+    const container = new Container();
+
+    container.applyMiddleware(debugMiddleware);
+    container.bind(TestService).toSelf();
+    container.get(TestService);
+
+    expect(consoleLogMock).not.toHaveBeenCalledWith(
+      messagesMap.providerRequested(TestService, container.id)
+    );
+  });
+
+  it("Should log a message for each provider resolved when logLevel is 'debug'.", () => {
     @injectable()
     class TestService1 {}
 
     @injectable()
     class TestService2 {}
 
-    inversilandOptions.debug = true;
+    inversilandOptions.logLevel = "debug";
 
     const consoleLogMock = jest.spyOn(console, "log").mockImplementation();
     const container = new Container();
@@ -59,10 +77,10 @@ describe("debugMiddleware", () => {
     container.get(TestService1);
     container.get(TestService2);
 
-    expect(consoleLogMock).toBeCalledWith(
+    expect(consoleLogMock).toHaveBeenCalledWith(
       messagesMap.providerRequested(TestService1, container.id)
     );
-    expect(consoleLogMock).toBeCalledWith(
+    expect(consoleLogMock).toHaveBeenCalledWith(
       messagesMap.providerRequested(TestService2, container.id)
     );
   });